refactor(frontend): extract blobDownload; deduplicate blob fallback path
Item 1 - blobDownload was defined identically in both Review.tsx (for the
single-file JSON save) and datasetZip.ts (for the zip blob fallback).
Extracted to src/download.ts; both files now import from there.

Item 4 - downloadDatasetZip's blob fallback path was duplicating
buildDatasetZipBlob's chunk-collection logic inline. The fallback now
calls buildDatasetZipBlob directly, leaving one canonical path for
turning a BuildDatasetZipOptions into a Blob.
