+++ packages/frontend/src/download.ts
@@ -1,0 +1,11 @@
+/** Triggers a browser download of `blob` using a temporary object URL. */
+export function blobDownload(blob: Blob, filename: string): void {
+  const url = URL.createObjectURL(blob);
+  const a = document.createElement('a');
+  a.href = url;
+  a.download = filename;
+  document.body.appendChild(a);
+  a.click();
+  document.body.removeChild(a);
+  URL.revokeObjectURL(url);
+}
